# lib/api-zod/src/generated/types/createCommissionPaymentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BD Tech Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCommissionPaymentInput {
  dealId: string;
  component: string;
  competenceMonth: string;
  /** @nullable */
  installmentIndex?: number | null;
  amount: string;
  /** @nullable */
  recipientUserId?: string | null;
  /** @nullable */
  paidByDirectorAt?: string | null;
  /** @nullable */
  confirmedByUserAt?: string | null;
  /** @nullable */
  rejectedByUserAt?: string | null;
  /** @nullable */
  isTestData?: boolean | null;
}